helper_scripts/requirements.txt
psycopg2==2.9.3
python-dotenv==0.19.2
